move new class to package structure analogue to matsim core, remove unused imports
